+++ src/start.cli.ts
@@ -10,6 +10,7 @@ import {
     config,
     extractArgs,
     gitInfo,
+    gqlGen,
     readVersion,
     run,
     start,
@@ -83,6 +84,27 @@ yargs.completion("completion")
             process.exit(101);
         }
 
+    })
+    .command(["gql-gen <glob>", "gql <glob>"], "Generate TypeScript code for GraphQL", ya => {
+        return ya
+            .option("change-dir", {
+                alias: "C",
+                default: process.cwd(),
+                describe: "Path to automation client project",
+                type: "string",
+            })
+            .option("install", {
+                default: true,
+                describe: "Run 'npm install' before starting/compiling",
+                type: "boolean",
+            });
+    }, argv => {
+        const glob = argv.glob;
+        gqlGen(argv["change-dir"], argv.glob, argv.install)
+            .then(status => process.exit(status), err => {
+                console.error(`${Package}: Unhandled Error: ${err.message}`);
+                process.exit(101);
+            });
     })
     .command("git", "Create a git-info.json file", ya => {
         return ya